requirements.txt
pillow
pyautogui
opencv-python
